src/recommendation_engine.py
```python
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from sklearn.feature_extraction.text import CountVectorizer

# Importing user profile and product data
from user_profile import userProfileData
from product_search import productData

def generateRecommendations():
    # Extracting features from product data
    features = ['name', 'category', 'description', 'brand']
    for feature in features:
        productData[feature] = productData[feature].fillna('')

    # Creating a combined features column
    def combine_features(row):
        return row['name'] + ' ' + row['category'] + ' ' + row['description'] + ' ' + row['brand']
    productData["combined_features"] = productData.apply(combine_features, axis=1)

    # Creating a count matrix from this new combined column
    cv = CountVectorizer()
    count_matrix = cv.fit_transform(productData["combined_features"])

    # Computing the Cosine Similarity based on the count_matrix
    cosine_sim = cosine_similarity(count_matrix)

    # Getting the index of the product that matches the title
    def get_index_from_title(title):
        return productData[productData.title == title]["index"].values[0]

    # Getting the title of the product that matches the index
    def get_title_from_index(index):
        return productData[productData.index == index]["title"].values[0]

    # Getting the user's favorite product from their profile
    user_favorite_product = userProfileData['favorite_product']

    # Getting the index of this product from its title
    product_index = get_index_from_title(user_favorite_product)

    # Enumerating through all the similarity scores of the user's favorite product to get similar products
    similar_products = list(enumerate(cosine_sim[product_index]))

    # Sorting the list similar_products in descending order of the similarity score
    sorted_similar_products = sorted(similar_products, key=lambda x:x[1], reverse=True)

    # Printing titles of first 10 products
    i = 0
    for product in sorted_similar_products:
        print(get_title_from_index(product[0]))
        i += 1
        if i > 10:
            break
```